[IMP]pms_fastapi: Add DELETE endpoint for user image.
