+++ roomdoo_fastapi/models/__init__.py
@@ -1,4 +1,3 @@
 from . import res_config_settings
 from . import res_users
 from . import auth_jwt_validator
-from . import res_partner
